Improve posted query test and fix test description text
